fix editor scroll; move MobileView and related out of primitives
