+++ packages/databyss-ui/primitives/index.js
@@ -19,7 +19,4 @@ export { default as RichTextInput } from './Control/native/RichTextInput'
 export { default as TextInputView } from './Control/native/TextInputView'
 export { default as RawHtml } from './Text/RawHtml'
 export { default as Switch } from './Control/native/Switch'
-export { default as ViewHeader } from './Mobile/ViewHeader'
-export { default as ViewHeaderNavItem } from './Mobile/ViewHeaderNavItem'
-export { default as MobileView } from './Mobile/MobileView'
 export { default as styled } from './styled'
